lizo/requirements.txt
openai>=1.0.0
edge-tts>=6.1.0
SpeechRecognition>=3.10.0
PyAudio>=0.2.13
flask>=3.0.0
flask-cors>=4.0.0
pyserial>=3.5
pyyaml>=6.0
